Parameterize database host per PR
